Check location of config.rb.
Because there are at least two use cases for Compass we should check where the config.rb file is. First we check whether config.rb is in the sublime project's root. If it is not there, then we check to see if the config.rb file is in the same directory as the .sass or .scss file.